Test refactor
These changes add a bit of duplication in favour of better readaility.
