+++ spec/support/test_helpers/candidate_helper.rb
@@ -1,13 +1,11 @@
 module CandidateHelper
-  def given_i_am_signed_in
+  def create_and_sign_in_candidate
     candidate = create(:candidate)
+
     login_as(candidate)
   end
 
-  def and_i_filled_in_personal_details
-    scope = 'application_form.personal_details'
-    visit candidate_interface_personal_details_edit_path
-
+  def candidate_fills_in_personal_details(scope:)
     fill_in t('first_name.label', scope: scope), with: 'Lando'
     fill_in t('last_name.label', scope: scope), with: 'Calrissian'
 
@@ -23,15 +21,5 @@ module CandidateHelper
 
     choose 'Yes'
     fill_in t('english_main_language.yes_label', scope: scope), with: "I'm great at Galactic Basic so English is a piece of cake", match: :prefer_exact
-
-    click_button t('complete_form_button', scope: scope)
-  end
-
-  def and_i_visit_the_application_form_page
-    visit candidate_interface_application_form_path
-  end
-
-  def when_i_click_on_check_your_answers
-    click_link 'Check your answers before submitting'
   end
 end
